fix: On-Prem Orch deployment in environments requiring a proxy (#35)
Fixes on-prem deployment in environements that require a proxy to reach
the internet. When the config is written for Orch, any changes applied
to it, like applying the proxy, are overwritten in the subsequent
execution of the installer.

This change makes it so that the subsequent execution skips the download
so it doesn't override the modified configuration including potential
proxy settings.