Mark tests with intermittent failures as "Pending" (#66)
- Some tests are failing intermittently in VIP and causing problematic
merge-queue failures and delays. These need to be debugged to determine
if the intermittent errors are test failures or product logic failures.


### Description

Please include a summary of the changes and the related issue. List any
dependencies that are required for this change.

Fixes # (issue)

### Any Newly Introduced Dependencies

Please describe any newly introduced 3rd party dependencies in this
change. List their name, license information and how they are used in
the project.

### How Has This Been Tested?

Please describe the tests that you ran to verify your changes. Provide
instructions so we can reproduce. Please also list any relevant details
for your test configuration

### Checklist:

- [x] I agree to use the APACHE-2.0 license for my code changes
- [x] I have not introduced any 3rd party dependency changes
- [x] I have performed a self-review of my code

Co-authored-by: Matteo <teone@users.noreply.github.com>