Pass argo.clusterDomain to keycloak-instance Helm chart values
The keycloak.yaml template needs access to clusterDomain to configure
the external HTTPS hostname. This was failing because the value wasn't
being passed from the ArgoCD Application to the Helm chart.

Now passing argo.clusterDomain in the valuesObject so keycloak.yaml
can use it to template: https://keycloak.{{ .Values.argo.clusterDomain }}